requirements.txt
# PhenoHunter Scientific - Python Dependencies

# Core Scientific Computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
torch>=1.9.0
scikit-learn>=1.0.0

# Visualization
plotly>=5.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Data Handling
pyyaml>=5.4.0
openpyxl>=3.0.0

# Web Scraping (optional)
requests>=2.26.0
beautifulsoup4>=4.9.0
selenium>=3.141.0

# PDF Processing (optional)
pdfplumber>=0.5.0
tabula-py>=2.3.0

# Utilities
fuzzywuzzy>=0.18.0
python-levenshtein>=0.12.0

# Jupyter Support
ipywidgets>=7.6.0
notebook>=6.4.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Code Quality
black>=21.0
flake8>=3.9.0
mypy>=0.910
